+++ src/index.pug
@@ -6,16 +6,51 @@ block content
     .maincontent
         section.section.welcome
             h1.visually-hidden Личный сайт разработчика Леонидовой Екатерины  
-            a.welcome-btn.welcome__button(href="auth.html" title="Авторизация") Авторизоваться
-            .welcome__inner
-                .profile.header__profile
-                    .photo
-                        img.img(src="img/me.jpg" alt="Леонидова Екатерина")
-                    h2.profile__title Леонидова Екатерина
-                    p.profile__text Веб-разработчик
-                    include pug/socials.pug
-                +nav(0)
-
+            button.btn.welcome-btn.welcome__button(title="Авторизация" id="auth") Авторизоваться
+            .welcome__flipper.flipper
+                .welcome__inner.flipper__front
+                    .profile.header__profile
+                        .photo
+                            img.img(src="img/me.jpg" alt="Леонидова Екатерина")
+                        h2.profile__title Леонидова Екатерина
+                        p.profile__text Веб-разработчик
+                        include pug/socials.pug
+                    +nav(0)
+                .auth.flipper__back
+                    h2.section__subtitle.section__subtitle--white.auth__title Авторизация
+                    form.auth__form.form.form--contact(action="")
+                        .form__input
+                            span.form__pic
+                                svg.form__svg(width="17" height="17")
+                                    use(xlink:href="img/sprite.svg#login")
+                            label.visually-hidden(for="login") Логин
+                            input.input.input--half-round(type="text" name="login" placeholder="Логин" id="login") 
+                        .form__input
+                            span.form__pic
+                                svg.form__svg(width="17" height="17")
+                                    use(xlink:href="img/sprite.svg#password")
+                            label.visually-hidden(for="password") Пароль
+                            input.input.input--half-round(type="password" name="password" placeholder="Пароль" id="password")  
+                        .form__group
+                            label.form__checkbox
+                                input.form__hidden(type="checkbox" id="human" name="human" value="1" checked)
+                                span.form__switch.form__switch--checkbox 
+                                span.form__label Я человек 
+                        .form__group.form__group--last
+                            span.form__text А вы точно не робот?
+                            .form__radios
+                                label.form__radio
+                                    input.form__hidden(type="radio" name="robot" value="Да" checked)
+                                    span.form__switch.form__switch--radio
+                                    span.form__label.form__label--green  Да
+                                label.form__radio 
+                                    input.form__hidden(type="radio" name="robot" value="Не уверен")
+                                    span.form__switch.form__switch--radio
+                                    span.form__label.form__label--green  Не уверен
+                        .form__buttons
+                            button.btn.form__btn(id="to-main" title="Главная") На главную
+                            button.btn.form__btn(type="submit") Войти
                
             .copyright.welcome__copyright © Екатерина Леонидова | создано с любовью в LoftSchool | 2018
-
+block script
+    script(src='js/index.bundle.js')
