--- src/js/modules/map.js
@@ -1,23 +1,0 @@
-var map = {
-    // initMap() {
-    //     var uluru = {lat: 34.013016, lng: -118.491651};
-    //     var map = new google.maps.Map(document.getElementById('map'), {
-    //       zoom: 10,
-    //       center: uluru
-    //       // 34.002510, -118.458889
-    //       // 33.997616, -118.434189
-    //       // 34.013016, -118.491651
-
-    //     });
-    //     var marker = new google.maps.Marker({
-    //       position: uluru,
-    //       map: map
-    //     });
-    //   }
-    // hi() {
-    //     console.log('hi');
-    // }
-};
-
-// module.exports = initMap;
-module.exports = map;
