+++ src/js/works.js
@@ -1,0 +1,3 @@
+let ham = require ('./modules/hamburger');
+
+ham.init();
