+++ src/pug/base.pug
@@ -12,7 +12,7 @@ html
         block head
     body
         block content
-        script(src='js/script.min.js')
+        //- script(src='js/script.min.js')
         block script
         //- script(src='https://maps.googleapis.com/maps/api/js?key=AIzaSyCdiqpmw5rW0CewRcKKUTZCtx4ePiFkEBA&callback=initMap')
