+++ src/scss/blocks/about-me.scss
@@ -2,6 +2,7 @@
     width: 100%;
     min-height: 100%;
     height: 100%;
+    position: relative;
 }
 
 .about-me__main {
