+++ src/scss/blocks/auth.scss
@@ -1,5 +1,5 @@
 .auth {
-    width: rem(380px);
+    min-width: rem(380px);
     background-color: rgba(#000, 0.5);
     display: flex;
     flex-direction: column;
@@ -8,7 +8,7 @@
     padding-top: rem(30px);
 
     @include screen-sm {
-        width: rem(348px);
+        min-width: rem(348px);
     }
 }
 
@@ -16,6 +16,3 @@
     margin: rem(20px) auto;
 }
 
-.auth__form {
-    
-}
