+++ src/scss/blocks/blog.scss
@@ -3,6 +3,7 @@
     flex-direction: column;
     width: 100%;
     color: $text;
+    position: relative;
 }
 
 .blog__main {
