+++ src/scss/blocks/flipper.scss
@@ -1,0 +1,56 @@
+.flipper {
+    position: relative;
+    perspective: 600px;
+
+    &--flip {
+        .flipper__front {
+            z-index: 900;
+            // border-color: #eee;
+            // background: #333;
+            // box-shadow: 0 15px 50px rgba(0,0,0,0.2);
+        
+            transform: rotateY(179deg);
+        }
+
+        .flipper__back {
+            z-index: 1000;
+            // background: #80868d;
+            transform: rotateX(0) rotateY(0);
+    
+            //  box-shadow: 0 15px 50px rgba(0,0,0,0.2);
+        }
+        
+    }
+
+
+}
+
+.flipper__front {
+    position: absolute;
+    top: 0;
+    left: 0;
+    z-index: 900;
+
+    transform: rotateX(0) rotateY(0);
+    transform-style: preserve-3d;
+    backface-visibility: hidden;
+    transition: all .4s ease-in-out;
+}
+
+.flipper__back {
+    position: absolute;
+    top: 0;
+    left: 0;
+    z-index: 800;
+    // width: inherit;
+    // height: inherit;
+    // border: 1px solid #ccc;
+    // background: #333;
+    // text-shadow: 1px 1px 1px rgba(0,0,0,0.6);
+
+    transform: rotateY(-179deg);
+    transform-style: preserve-3d;
+    backface-visibility: hidden;
+    transition: all .4s ease-in-out;
+  }
+
