+++ src/scss/blocks/fullscreen.scss
@@ -1,0 +1,55 @@
+.fullscreen {
+    width: 100%;
+    height: 100vh;
+    position: fixed;
+    top: 0;
+    left: 0;
+    z-index: 30;
+    background-color: $lightGreen;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    .nav__list {
+        display: flex;
+        flex-direction: column;
+        
+    }
+    
+    .nav__item {
+        flex-grow: 0;
+        background-color: transparent;
+        border-left: 0;
+        // align-items: center;
+        position: relative;
+          
+        &:hover {
+            background-color: transparent;
+            &::after {
+                content: '';
+                position: absolute;
+                bottom: 0;
+                left: 50%;
+                margin-left: rem(-45px);
+                background-color: #fff;
+                width: rem(90px);
+                height: 3px;    
+            }          
+        }
+    
+    }
+    
+    .nav__link {
+        display: block;
+        padding: 13px 15px;
+        @include sansus;
+        font-size: rem(49px);
+        color: #fff;
+        text-transform: uppercase;
+        width: 100%;
+        height: 100%;
+        text-align: center;
+
+    }
+    
+}
