+++ src/scss/blocks/hamburger-menu.scss
@@ -1,0 +1,71 @@
+.hamburger-menu
+{
+    position: relative;
+
+    display: flex;
+
+    width: 40px;
+    height: 35px;
+
+    cursor: pointer;
+    vertical-align: middle;
+
+    align-items: center;
+
+
+    &--close
+    {
+        .hamburger-menu__bars
+        {
+            transform: rotate(45deg);
+
+            &:before
+            {
+                transform: rotate(45deg);
+
+                opacity: 0;
+            }
+
+            &:after
+            {
+                transform: rotate(-90deg);
+            }
+        }
+    }
+}
+
+.hamburger-menu__bars
+{
+    &,
+    &:before,
+    &:after
+    {
+        width: 40px;
+        height: 5px;
+
+        transition: transform .3s;
+
+        border-radius: 3px;
+        background: #fff;
+    }
+
+    &:before,
+    &:after
+    {
+        position: absolute;
+
+        display: block;
+
+        content: '';
+    }
+
+    &:before
+    {
+        top: 0;
+    }
+
+    &:after
+    {
+        bottom: 0;
+    }
+}
